Add a restart policy to make our contianers come back when the server reboots.
This actually works! CoreOS periodically reboots to install updates,
which feels like a good thing to me, but I do want the contianers to
come back up when it does. The coreos docs recommended setting up a
full systemd unit and service for this, but some googling suggested an
even simpler solution: using the `--restart` parameter to docker run.

Previously we were scp-ing some config files to /tmp, which gets wiped
on reboot, so I moved those to /home/jeremyfleischman, which isn't a
very portable name. Maybe we can use `.` and `$PWD` as necessary to make
this more generic in the future, or maybe put things in a deterministic
place? `/etc/gatekeeper/...`?